Move NYI options to a different branch
I'm doing this to prepare the main branch for smaller releases
with just a couple features in each of them. I cannot do
everything in a single release and I don't want to leave
NYI options in, this will confuse users.
I will slowly implement everything and add these options back
into the codebase.

Wish me luck, it's a lot.